[Common] period 타입 Int여도 충분할 것 같아서 변경

+++ shared/src/main/kotlin/com/lovelessgeek/housemanager/shared/models/Task.kt
@@ -13,9 +13,9 @@ data class Task(
     val time: Date,
     val isComplete: Boolean = false,
     val isRepeat: Boolean = false,
-    val period: Long = PERIOD_NO_REPEAT
+    val period: Int = PERIOD_NO_REPEAT
 ) {
     companion object {
-        const val PERIOD_NO_REPEAT = 0L
+        const val PERIOD_NO_REPEAT = 0
     }
 }
